# api/api.yaml
openapi: 3.0.1  
info:
  version: 1.0.0
  title: Hacker News
  description: "This is the Open API documentation for the REST API of our beloved application **Hacker News** deployed at <https://hacker-news-fib.herokuapp.com/>. <br>All operations are executable. Only one operation requires authentication: `deleteTweet`. In this case, you must **Authorize** your request by providing the api_key vaule you got when you created the tweet."
servers:
  - url: 'http://127.0.0.1:3000/api/v1/'
paths:
...
  /posts:
    get:
      tags:
        - posts
      summary: Finds Posts
      description: Return the Posts
      operationId: getPosts
      parameters:
      - name: user_id
        in: query
        description: ID of the user
        schema:
          type: integer
          format: int64
      - name: order_by
        in: query
        description: Type of ordering
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostList'
        400:
          description: 'Error: Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error400Order'
        404:
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorUser:
                  $ref: '#/components/examples/error404User'
                  
    post:
      tags:
        - posts
      summary: Creates a new Post
      description: Creates a new Post and returns its info
      operationId: newPost
      requestBody:
        description: Provide the title & content of the new Post
        content:
          application/json:
            schema:
              required:
              - title
              type: object
              properties:
                title:
                  type: string
                  minLength: 4
                  example: 'My name'
                url:
                  type: string
                  format: url
                  example: 'https://www.github.com'
                text:
                  type: string
                  minLength: 4
                  example: 'My text'
        required: true
      responses:
        201:
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddPostResult'
        400:
          description: 'Error: Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
        401:
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        409:
          description: 'Error: Conflict'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error409'
      security:
        - ApiKeyAuth: []

  ### /POSTS/:ID
...
  /posts/{postId}/vote:
    post:
      tags:
        - posts
        - votes
      summary: Vote a Post
      description: Vote a Post
      operationId: votePost
      parameters:
      - name: postId
        in: path
        description: ID of Post to return
        required: true
        schema:
          type: integer
          format: int64
      responses:
        201:
          description: Vote done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        401:
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        404:
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404Post'
        409:
          description: 'Error: Conflict'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error409Vote'
      security:
        - ApiKeyAuth: []
    delete:
      tags:
        - posts
        - votes
      summary: Unvote a Post
      description: Unvote a Post
      operationId: unvotePost
      parameters:
      - name: postId
        in: path
        description: ID of Post to return
        required: true
        schema:
          type: integer
          format: int64
      responses:
        204:
          description: Unvote done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        401:
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        403:
          description: 'Error: Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error403'
        404:
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404Post'
        409:
          description: 'Error: Conflict'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error409Unvote'
      security:
        - ApiKeyAuth: []    
...
  /comments/{commentId}/vote:
    post:
      tags:
        - votes
        - comments
      summary: Vote a Comment
      description: Vote a Comment
      operationId: voteComment
      parameters:
      - name: commentId
        in: path
        description: ID of Comment to vote
        required: true
        schema:
          type: integer
          format: int64
      responses:
        201:
          description: Vote done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        401:
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        404:
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404Comment'
        409:
          description: 'Error: Conflict'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error409Vote'
      security:
        - ApiKeyAuth: []
    delete:
      tags:
        - comments
        - votes
      summary: Unvote a Comment
      description: Unvote a Comment
      operationId: unvoteComment
      parameters:
      - name: commentId
        in: path
        description: ID of Comment to unvote
        required: true
        schema:
          type: integer
          format: int64
      responses:
        204:
          description: Unvote done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        401:
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        403:
          description: 'Error: Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error403'
        404:
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404Post'
        409:
          description: 'Error: Conflict'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error409Unvote'
      security:
        - ApiKeyAuth: []
  
  ### REPLIES
...
components:
  schemas:
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        title:
          type: string
          minLength: 4
          example: 'Post Title'
        url:
          type: string
          format: url
          example: 'https::www.github.com'
        text:
          type: string
          minLength: 4
          example: 'Post text'
        points:
          type: integer
          format: int64
          example: 5
        userId:
          type: integer
          format: int64
          example: 1
        created_at:
          type: string
          format: date-time
    PostList:
      type: array
      items:
        $ref: '#/components/schemas/Post'
...
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        text:
          type: string
          minLength: 4
          example: 'Comment Text'
        votes:
          type: integer
          format: int64
          example: 5
        userId:
          type: integer
          format: int64
          example: 3
        postId:
          type: integer
          format: int64
          example: 1
        created_at:
          type: string
          format: date-time
...
    AddPostResult:
      type: object
      properties:
        new_post:
          $ref: '#/components/schemas/Post'
...
    ErrorResult:
      type: object
      properties:
        status:
          type: integer
          enum: [400, 401, 403, 404, 409]
        error: 
          type: string
          enum: ["Bad Request", "Unauthorized", "Forbidden", "Not Found", "Conflict"]
        message: 
          type: string
  examples:
    error400:
      value:
        status: 400
        error: "Bad Request"
        message: "Content is empty"
    error400Order:
      value:
        status: 400
        error: "Bad Request"
        message: "Ordering not valid"
    error401:
      value:
        status: 401
        error: "Unauthorized"
        message: "You provided no api key (X-API-KEY Header)"
    error403:
      value:
        status: 403
        error: "Forbidden"
        message: "Your api key (X-API-KEY Header) is not valid"
    error404Comment:
      value:
        status: 404
        error: "Not Found"
        message: "No Comment with that ID"
...
    error404Post:
      value:
        status: 404
        error: "Not Found"
        message: "No Post with that ID"
    error404User:
      value:
        status: 404
        error: "Not Found"
        message: "No User with that ID"
    error409:
      value:
        status: 409
        error: "Conflict"
        message: "Post with the same url already created"
        post: 
          $ref: '#/components/schemas/Post'
...
    error409Vote:
      value:
        status: 409
        error: "Conflict"
        message: "Vote to the same Contribution already created"
    error409Unvote:
      value:
        status: 409
        error: "Conflict"
        message: "No Vote to the Contribution exists"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: X-API-KEY
      in: header